♻️ refactor: extract comment component
